Add SQLAlchemy database integration

+++ app/schemas.py
@@ -1,8 +1,9 @@
+from decimal import Decimal
 from pydantic import BaseModel, Field, field_validator
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=150)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=250)
 
     @field_validator("amount")
@@ -21,7 +22,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=50)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator("name")
     def check_name(cls, v: str) -> str:
@@ -31,7 +32,7 @@ class CreateWalletRequest(BaseModel):
         return v
 
     @field_validator("initial_balance")
-    def check_ib(cls, v: float) -> float:
+    def check_ib(cls, v: Decimal) -> Decimal:
         if v < 0:
             raise ValueError("Amount cant be negative")
         return v
